.git壊してからの更新分

+++ Game/Common/Camera/PlayerCamera/PlayerCamera.cpp
@@ -93,8 +93,8 @@ void PlayerCamera::Update(float deltaTime)
 	auto gamepad = GetCommonResources()->GetGamePadTracker()->GetLastState();
 	if (gamepad.IsConnected())
 	{
-		mouseX = gamepad.thumbSticks.rightX * 15.0f;
-		mouseY = -gamepad.thumbSticks.rightY * 10.0f;
+		mouseX = static_cast<int>(gamepad.thumbSticks.rightX * 15.0f);
+		mouseY = static_cast<int>(-gamepad.thumbSticks.rightY * 10.0f);
 	}
 	else
 	{
